style: api/posts 폴더 내 파일명 수정
