+++ src/api/posts/readPostDetail.ts
@@ -1,6 +1,6 @@
 import axios from "@api/axiosInstance";
 
-export const fetchPostDetail = async (id: string) => {
+export const readPostDetail = async (id: string) => {
   const response = await axios.get(`/posts/${id}`);
   return response.data;
 };
